Added URL checking
As input validation, the program checks if the URL given by the user is reachable or even existent. For that I used the module requests and the declaration r = requests.get(URL) to get a status code from the website. Then r.status_code returns the status code of the website, which if it exists or is responsive, will be 200, thus progressing with the program. If not, the program will output the problem and once again output the proper use of the program in the command line.