Changed order of declarations due to index error
Had to separate directory and URL declarations due to an index error that appeared when setting the variable for "my_url" in the case of there not being a second command line argument. Also moved the "user_dir" declaration under the "--help" command line identifier to make code easier to understand.